requirements.txt
playwright
py-esprima
PyYAML
requests
jsbeautifier
google-generativeai
jinja2